+++ src/modules/scenegraph/SceneGraphNodeProperties.h
@@ -27,6 +27,12 @@ static constexpr const char *PropCamFov = "cam_fov";
 static constexpr const char *PropCamFrustum = "frustum";
 static constexpr const char *PropParent = "parent";
 static constexpr const char *PropParentUUID = "parent-uuid";
+static constexpr const char *PropSunIntensity = "sun_intensity";
+static constexpr const char *PropSunArea = "sun_area";
+static constexpr const char *PropSunElevation = "sun_elevation";
+static constexpr const char *PropSunAzimuth = "sun_azimuth";
+static constexpr const char *PropSunColor = "sun_color";
+static constexpr const char *PropSunDisk = "sun_disk";
 
 using SceneGraphNodeProperties = core::DynamicStringMap<core::String>;